World Bank Capacity data set 2010-2016

Statistical Capacity Indicators provides information on various aspects of national statistical systems of developing countries, including an overall country-level statistical capacity indicator.

data(cap)

Format

An object of class "data.table".

Source

World Bank Capacity archive

Examples

data(cap)
str(cap)

#> Classes ‘data.table’ and 'data.frame':	1085 obs. of  31 variables:
#>  $ ISO                                                                                       : chr  "AFG" "AFG" "AFG" "AFG" ...
#>  $ YEAR                                                                                      : num  2010 2011 2012 2013 2014 ...
#>  $ CAP_ACCESS_TO_WATER_5_51_01_09_WATER_                                                     : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_AGRICULTURAL_CENSUS_3_01_04_01_AGCEN_                                                 : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ CAP_BALANCE_OF_PAYMENTS_MANUAL_IN_USE_2_04_01_01_EXCNCPT_                                 : num  0 0 0 1 1 1 1 1 1 1 ...
#>  $ CAP_CHILD_IMMUNIZATION_5_51_01_04_IMMUN_                                                  : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_CHILD_MALNUTRITION_5_51_01_02_MALNUT_                                                 : num  0.667 0.333 0.333 0.333 0.333 ...
#>  $ CAP_CHILD_MORTALITY_5_51_01_03_MORTAL_                                                    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_CONSUMER_PRICE_INDEX_BASE_YEAR_2_01_03_01_PRCPBASE_                                   : num  0 0 0 0 0 0 0 1 1 1 ...
#>  $ CAP_EXTERNAL_DEBT_REPORTING_STATUS_5_04_01_01_EXDEBT_                                     : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_GENDER_EQUALITY_IN_EDUCATION_5_51_01_07_GENDER_                                       : num  1 1 1 1 1 1 1 0 0 0 ...
#>  $ CAP_GOVERNMENT_FINANCE_ACCOUNTING_3_02_01_02_FSCOV_                                       : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_HEALTH_SURVEY_5_13_01_01_HLTHSURV_                                                    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_HIV_AIDS_5_51_01_05_HIV_                                                              : num  1 1 1 1 1 1 1 0 0 0 ...
#>  $ CAP_IMPORT_AND_EXPORT_PRICE_INDEXES_5_04_01_02_IMPEXP_                                    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ CAP_INCOME_POVERTY_5_51_01_01_POVERTY_                                                    : num  0 0 0 0 0 0 0 1 1 1 ...
#>  $ CAP_INDUSTRIAL_PRODUCTION_INDEX_5_01_01_01_INDUST_                                        : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ CAP_MATERNAL_HEALTH_5_51_01_06_MATERN_                                                    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_METHODOLOGY_ASSESSMENT_OF_STATISTICAL_CAPACITY_SCALE_0_100_IQ_SCI_MTHD_               : num  40 40 40 50 50 40 40 60 60 70 ...
#>  $ CAP_NATIONAL_ACCOUNTS_BASE_YEAR_2_01_01_02_NABASE_                                        : num  1 1 1 1 1 0 0 1 1 1 ...
#>  $ CAP_NATIONAL_IMMUNIZATION_COVERAGE_5_13_01_01_WHO_                                        : num  1 1 0 0 0 0 0 1 1 1 ...
#>  $ CAP_PER_CAPITA_GDP_GROWTH_5_51_01_10_GDP_                                                 : num  0.667 0.667 0.667 0.667 1 ...
#>  $ CAP_PERIODICITY_AND_TIMELINESS_ASSESSMENT_OF_STATISTICAL_CAPACITY_SCALE_0_100_IQ_SCI_PRDC_: num  76.7 73.3 70 70 73.3 ...
#>  $ CAP_POPULATION_CENSUS_3_11_01_01_POPCEN_                                                  : num  0 0 0 0 0 0 0 1 1 1 ...
#>  $ CAP_POVERTY_SURVEY_5_14_01_01_POVSURV_                                                    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CAP_PRIMARY_COMPLETION_5_51_01_08_PRIMCOMP_                                               : num  0.333 0.333 0 0 0 ...
#>  $ CAP_SOURCE_DATA_ASSESSMENT_OF_STATISTICAL_CAPACITY_SCALE_0_100_IQ_SCI_SRCE_               : num  40 40 40 40 40 40 40 80 80 80 ...
#>  $ CAP_SPECIAL_DATA_DISSEMINATION_STANDARD_5_21_01_01_SDDS_                                  : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ CAP_STATISTICAL_CAPACITY_SCORE_OVERALL_AVERAGE_IQ_SCI_OVRL_                               : num  52.2 51.1 50 53.3 54.4 ...
#>  $ CAP_UNESCO_REPORTING_5_12_01_01_UNESCO_                                                   : num  0 0 1 1 1 1 1 0 0 1 ...
#>  $ CAP_VITAL_REGISTRATION_SYSTEM_COVERAGE_3_11_01_03_POPREG_                                 : num  0 0 0 0 0 0 0 1 1 1 ...
#>  - attr(*, ".internal.selfref")=<externalptr>