Check if a submission to the DataChallenge has the right format.

Your submission should have the following columns

"TEAM_ID"a unique team ID (character string)

"SUBMISSION_ID"a unique ID of this subission (character string).

"ISO"country code of the country for which you are making a submission. You need to make a prediction for 66 countries.

"YEAR"year for which you are making a submission. This should be either 2015 or 2016 (integer).

"PREDICTION"Predicted number of epidemic outbreaks in this country and this year. This can be real value.

This function will print a warning if there are any issues with the format of your submission.

dc.check.submission(ev)

Arguments

ev
data.table with specific columns, see below. and predictions in the PREDICTION rows and

Value

TRUE or FALSE