World Bank Development Indicators data set 2010-2016

World Development Indicators (WDI) is the primary World Bank collection of development indicators, compiled from officially recognized international sources. It presents the most current and accurate global development data available, and includes national, regional and global estimates.

data(devi)

Format

An object of class "data.table".

Source

World Bank WDI archive

Examples

data(devi)
str(devi)

#> Classes ‘data.table’ and 'data.frame':	1841 obs. of  23 variables:
#>  $ ISO                                                                                                           : chr  "vree" "vree" "vree" "vree" ...
#>  $ YEAR                                                                                                          : num  2010 2011 2012 2013 2014 ...
#>  $ DEV_ACCESS_TO_CLEAN_FUELS_AND_TECHNOLOGIES_FOR_COOKING_OF_POPULATION_EG_CFT_ACCS_ZS_                          : num  19 18.5 18.1 17.7 17.3 ...
#>  $ DEV_ACCESS_TO_ELECTRICITY_OF_POPULATION_EG_ELC_ACCS_ZS_                                                       : num  42.7 61.5 69.1 75.2 89.5 ...
#>  $ DEV_ACCESS_TO_ELECTRICITY_RURAL_OF_RURAL_POPULATION_EG_ELC_ACCS_RU_ZS_                                        : num  32.4 55 63.8 70.1 87.8 ...
#>  $ DEV_ACCESS_TO_ELECTRICITY_URBAN_OF_URBAN_POPULATION_EG_ELC_ACCS_UR_ZS_                                        : num  82.8 81 95 89.6 98.7 ...
#>  $ DEV_ACCOUNT_AGE_15_TS_WP_TIME_10_1_                                                                           : num  NA 9.01 NA NA 9.96 ...
#>  $ DEV_ACCOUNT_AT_A_FINANCIAL_INSTITUTION_AGE_15_WP_TIME_01_1_                                                   : num  NA 9.01 NA NA 9.96 ...
#>  $ DEV_ACCOUNT_AT_A_FINANCIAL_INSTITUTION_FEMALE_AGE_15_WP_TIME_01_3_                                            : num  NA 2.62 NA NA 3.81 ...
#>  $ DEV_ACCOUNT_AT_A_FINANCIAL_INSTITUTION_INCOME_POOREST_40_AGES_15_WP_TIME_01_8_                                : num  NA 0.863 NA NA 6.619 ...
#>  $ DEV_ACCOUNT_AT_A_FINANCIAL_INSTITUTION_INCOME_RICHEST_60_AGES_15_WP_TIME_01_9_                                : num  NA 14.4 NA NA 12.2 ...
#>  $ DEV_ACCOUNT_AT_A_FINANCIAL_INSTITUTION_MALE_AGE_15_WP_TIME_01_2_                                              : num  NA 15.4 NA NA 15.8 ...
#>  $ DEV_ACCOUNT_FEMALE_AGE_15_TS_WP_TIME_10_3_                                                                    : num  NA 2.62 NA NA 3.81 ...
#>  $ DEV_ACCOUNT_INCOME_POOREST_40_AGES_15_TS_WP_TIME_10_8_                                                        : num  NA 0.863 NA NA 6.619 ...
#>  $ DEV_ACCOUNT_INCOME_RICHEST_60_AGES_15_TS_WP_TIME_10_9_                                                        : num  NA 14.4 NA NA 12.2 ...
#>  $ DEV_ACCOUNT_MALE_AGE_15_TS_WP_TIME_10_2_                                                                      : num  NA 15.4 NA NA 15.8 ...
#>  $ DEV_ACCOUNT_OLDER_ADULTS_AGES_25_TS_WP_TIME_10_5_                                                             : num  NA 10.5 NA NA 11.5 ...
#>  $ DEV_ACCOUNT_PRIMARY_EDUCATION_OR_LESS_AGES_15_TS_WP_TIME_10_6_                                                : num  NA 5.32 NA NA 4.83 ...
#>  $ DEV_ACCOUNT_SECONDARY_EDUCATION_OR_MORE_AGES_15_TS_WP_TIME_10_7_                                              : num  NA 30 NA NA 22.6 ...
#>  $ DEV_ACCOUNT_YOUNG_ADULTS_AGES_15_24_TS_WP_TIME_10_4_                                                          : num  NA 6.3 NA NA 7.45 ...
#>  $ DEV_ADEQUACY_OF_SOCIAL_INSURANCE_PROGRAMS_OF_TOTAL_WELFARE_OF_BENEFICIARY_HOUSEHOLDS_PER_SI_ALLSI_ADQ_POP_TOT_: num  NA NA NA NA NA ...
#>  $ DEV_X2005_PPP_CONVERSION_FACTOR_PRIVATE_CONSUMPTION_LCU_PER_INTERNATIONAL_PA_NUS_PRVT_PP_05_                  : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ DEV_X2005_PPP_CONVERSION_FACTOR_GDP_LCU_PER_INTERNATIONAL_PA_NUS_PPP_05_                                      : num  NA NA NA NA NA NA NA NA NA NA ...
#>  - attr(*, ".internal.selfref")=<externalptr>