IHR data set 2010-2016

As part of the International Health Regulations monitoring framework, countries are requested to self assess their capacity in responding to epidemic disasters each year. Self-assessments range from 0 to 100, and are sub-divided into 11 categories.

data(ihr)

Format

An object of class "data.table".

Source

IHR Data Archive

Examples

data(ihr)
str(ihr)

#> Classes ‘data.table’ and 'data.frame':	976 obs. of  13 variables:
#>  $ ISO                   : chr  "AFG" "AFG" "AFG" "AFG" ...
#>  $ YEAR                  : int  2014 2011 2012 2013 2016 2015 2010 2014 2011 2012 ...
#>  $ IHR_LEGISLATION       : num  50 0 0 0 0 0 50 100 0 0 ...
#>  $ IHR_COORDINATION      : num  63 20 36 46 30 36 75 100 20 53 ...
#>  $ IHR_SURVEILLANCE      : num  90 80 80 75 95 75 96 85 0 35 ...
#>  $ IHR_RESPONSE          : num  58 43 47 40 87.5 52 53 94 28 35 ...
#>  $ IHR_PREPAREDNESS      : num  26 0 0 10 28.3 ...
#>  $ IHR_RISK_COMMUNICATION: num  71 43 43 57 57.1 ...
#>  $ IHR_HUMAN_RESOURCES   : num  20 25 20 40 40 40 16 60 0 0 ...
#>  $ IHR_LABORATORY        : num  53 40 26 45 75.8 ...
#>  $ IHR_POINTS_OF_ENTRY   : num  15 0 3 3 12.1 ...
#>  $ IHR_ZOONOSIS          : num  67 44 56 67 33.3 ...
#>  $ IHR_FOOD_SAFETY       : num  20 25 13 20 6.67 ...
#>  - attr(*, ".internal.selfref")=<externalptr>